+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Log the `Content-Type`, `Origin` and `Sec-Fetch-*`, `X-Forwarded-For` header values for each request.
 * Added request completion time logging.
 * For `POST` requests, log the size of the request body.
+* Remove the certificate subject common name (CN) from the log's.
 
 ## [1.6.0] Improved Key-Capsule Expiry Logic & secp265 Support (2025-12-01)